requirements.txt
django
djangorestframework
django-cors-headers
django-currentuser
django-summernote
pillow